[#98]rename: authEmail -> user 폴더 이동
